Factor SimulatedUser behavior out into a helper class so we can have multiple versions.
